Added support to both the --features and --hotfixes flags

+++ src/index.js
@@ -1,6 +1,11 @@
 import { program } from "commander";
-import { deleteAllGitBranches } from "./gitCommands.js";
-// git-branch-clean clean --except -x --all -a
+import {
+  deleteAllGitBranches,
+  deleteAllGitFeatureBranches,
+  deleteAllGitHotFixBranches,
+} from "./gitCommands.js";
+// git-branch-clean clean --except -x --all -a -c --clean
+// we won't support the except flag on the first release
 program
   .option(
     "-x, --except <string>",
@@ -8,6 +13,11 @@ program
     (value) => value.split(/,/)
   )
   .option("-a, --all", "remove all branches except master/main", false)
+  .option(
+    "-c, --clean",
+    "remove all branches except master/main/develop",
+    false
+  )
   .option(
     "-f, --features",
     "removes all branches that start with feature/*",
@@ -23,7 +33,16 @@ async function cli() {
   program.parse();
   const options = program.opts();
   try {
-    await deleteAllGitBranches();
+    if (options.all) {
+      await deleteAllGitBranches();
+      return;
+    }
+    if (options.features) {
+      await deleteAllGitFeatureBranches();
+    }
+    if (options.hotfixes) {
+      await deleteAllGitHotFixBranches();
+    }
   } catch (error) {
     console.log(error.message);
   }
